KS-1-4 Variable type + correction

+++ src/main/kotlin/lesson_1/lesson_1_task_4.kt
@@ -5,8 +5,9 @@ fun main() {
     val age: Byte = 27
     val partOfDay: Double = 0.075
     val seconds: Short = 6480
-    val partOfYear: Float = 2.0547945205479453E-4f
+    val partOfYear: Double = 2.0547945205479453E-4
     val apogee: Int = 327000
+
     println(length)
     println(age)
     println(partOfDay)
